move MDI C style library calls to separate files
