+++ src/USER-MDI/fix_mdi_engine.cpp
@@ -17,6 +17,7 @@
 ------------------------------------------------------------------------- */
 
 #include "fix_mdi_engine.h"
+#include "library_mdi.h"
 
 #include "atom.h"
 #include "comm.h"
